+++ pageindex/__init__.py
@@ -1,6 +1,7 @@
 """PageIndex SDK."""
 from typing import TYPE_CHECKING as _TYPE_CHECKING
 
+from .chat_stream import ChatStream
 from .client import PageIndexClient, PageIndexCloudClient, PageIndexLocalClient
 from .errors import PageIndexAPIError
 from .types import (ChatConfig, ChatProcessOptions, CloudIndexConfig,
@@ -8,7 +9,6 @@ from .types import (ChatConfig, ChatProcessOptions, CloudIndexConfig,
 
 if _TYPE_CHECKING:
     from .flash import page_index_flash
-    from .local_chat import ChatStream
     from .page_index_classic import page_index, page_index_main
     from .page_index_md import md_to_tree
     from .tree_optimize import optimize_tree
@@ -23,15 +23,14 @@ __all__ = [
 ]
 
 _LAZY = {
-    "ChatStream": ".local_chat",
     "page_index_flash": ".flash",
     "optimize_tree": ".tree_optimize",
     "md_to_tree": ".page_index_md",
 }
-_SUBMODULES = {"agent_tools", "client", "cloud_api", "errors", "flash",
-               "integrations", "local_api", "local_chat", "local_store",
-               "mcp_bridge", "page_index_classic", "page_index_md",
-               "tree_optimize", "types", "utils"}
+_SUBMODULES = {"agent_tools", "chat_stream", "client", "cloud_api", "errors",
+               "flash", "integrations", "local_api", "local_chat",
+               "local_store", "mcp_bridge", "page_index_classic",
+               "page_index_md", "tree_optimize", "types", "utils"}
 
 
 def __getattr__(name):
