+++ pageindex/local_chat.py
@@ -13,6 +13,7 @@ import uuid
 from typing import Any, Iterator, Mapping, Optional, Union
 
 from .agent_tools import _base_instructions, doc_targeting_block
+from .chat_stream import ChatStream
 from .errors import PageIndexAPIError
 
 CHAT_HEADER = (
@@ -772,68 +773,6 @@ def _weave(events, options) -> Iterator[str]:
         close = getattr(events, "close", None)
         if close is not None:
             close()  # cancel the underlying run on abandonment
-
-
-class ChatStream:
-    """chat(stream=True)'s stream: iterate it for the answer text pieces
-    (with show_process, the woven display); read ``.events`` instead for
-    the typed process event dicts. One underlying run — consume exactly
-    one view; call chat() again for the other."""
-
-    def __init__(self, text, events):
-        self._text = text      # () -> Iterator[str]
-        self._events = events  # () -> Iterator[dict], or the refusal text
-        self._view: Optional[str] = None
-        self._it: Any = None
-        self._closed = False
-
-    def _claim(self, view: str) -> None:
-        if self._view is not None and self._view != view:
-            raise PageIndexAPIError(
-                f"This chat stream is being consumed as {self._view}; one "
-                "run serves one view — call chat() again for the other.")
-        self._view = view
-
-    def __iter__(self) -> "ChatStream":
-        return self
-
-    def __next__(self) -> str:
-        self._claim("text")
-        if self._it is None:
-            if self._closed:
-                raise StopIteration
-            self._it = self._text()
-        return next(self._it)
-
-    @property
-    def events(self) -> Iterator[dict]:
-        """The run as typed event dicts: {"type": "thinking"|"answer",
-        "delta": ...}, {"type": "tool_call", "call_id", "name",
-        "arguments"}, {"type": "tool_result", "call_id", "name",
-        "output"} — full data, never clipped. Consuming — not merely
-        reading the attribute — claims the view, so debugger panes and
-        getattr probing stay side-effect free."""
-        def consume():
-            if isinstance(self._events, str):
-                raise PageIndexAPIError(self._events)
-            self._claim("events")
-            if self._it is None:
-                if self._closed:
-                    return
-                self._it = self._events()
-            # no `yield from`: a dropped handle must not close the run
-            for ev in self._it:
-                yield ev
-        return consume()
-
-    def close(self) -> None:
-        """Stop the run: closes the open view, and the stream is dead
-        afterwards, like a closed generator (own-model chat: a run never
-        consumed never starts)."""
-        self._closed = True
-        close = getattr(self._it, "close", None)
-        if close is not None:
-            close()
 
 
 def _cloud_chunk_events(chunks) -> Iterator[dict]:
